Fix circular dependency in frontend target_system module
